Make the expiration thread a daemon thread
I might be missing something, but it seems to me that the expiration thread should be a daemon.

Right now, it is difficult for my application to go through an orderly shutdown because the `executorN` threads survive a call to `shutdown` on the `Setup` object, and because this thread survives. The former I can get around by calling `shutdown` on the global pool (although this is not an ideal solution if I have several `Setup` instances), but the latter thread I cannot shut down.